chatbots/white-label-chatbot/requirements.txt
fastapi
langchain
langchain-anthropic
langchain-cohere
langchain-openai
lunary
psycopg2
pydantic
python-dotenv
requests
SQLAlchemy
tokenizers
uvicorn
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
langchainhub
cohere
alembic
psycopg2-binary
pandas
pillow